Display the research-interests visibility next to the text
This change adds text to the value to say who can view it to help make it more transparent, especially for those who have already set research interests before we added the visibility option. This text means you don't have to look at the form or the profile page to check.

This page is getting quite crowded with various options.

Refs #1208