Allow users to connect their Community Slack account
This change adds a page to allow users to use the 'Sign in with Slack' flow to connect their accounts. This process is non-reversible: revoking the authorisation on Slack's side doesn't stop PREreview from reading their profile information (as it uses a Bot Token). Note this is behind a feature flag.

Refs #1251, https://api.slack.com/authentication/sign-in-with-slack, https://api.slack.com/authentication/token-types