Ensure the verification email looks right
HTML emails are notoriously hard to produce. This change adds a snapshot to check if it looks ok. This test isn't the same as ensuring it looks right inside an email client, but it's the best we can do.

I've had to change the link text as it contains a generated UUID and port number, which differ each time.

Refs #1307, https://explore.reallygoodemails.com/new-to-email-coding-heres-where-to-start-2494422f0bd4