Remove knowledge of the user and flash message from the homepage
This change introduces an idea we've talked about for quite a while: decoupling pages from hyper-ts.

It's a prerequisite to two things:
Supporting logging in from any page
Being able to nudge new users to their 'My details' page

We could achieve both without this change, but it would mean propagating knowledge of both everywhere. So, it seems like a sensible time to introduce this change.

Refs #1396