Refresh the Zenodo data when publishing a PREreview and adding an author
This change tries to ensure that if the user looks at the reviews for the preprint or at the review directly after pressing publish, they see their changes.

This behaviour is a hack and only works for the instance that received the publish request. It shows we should find a better way to cache PREreview data locally and invalidate entries across all the instances.

Refs #1513