Calm down sending author invites
We've seen a second case of many invites (18; the first case was 8) failing due to Redis timeouts. It might be related to ioredis's autopipeling and that writes to a replica are forwarded to the primary. This change tries to mitigate the problem by handling each invite one at a time. Sending many invites through a replica might now be noticeably slow, but sending 18 is unusual.

Refs https://github.com/redis/ioredis/blob/ec42c82ceab1957db00c5175dfe37348f1856a93/README.md#autopipelining, https://fly.io/docs/reference/redis/#writing-to-replica-regions