Ensure the conversions between ISO 639-1 and 639-3 are type-safe
This change makes it easier to work with the ISO 639-3 codes returned by Zenodo.

Note I had to replicate the mapping from the iso-639-3 library as TypeScript widens string value types. I've also had to exclude 'hbs' as the iso-639-1 library doesn't include the equivalent code 'sh'. As it's deprecated, this shouldn't be a problem.

Refs #1739, https://github.com/microsoft/TypeScript/issues/32063