Stop the server from crashing
When upgrading the http-proxy-middleware library I accidentally caused the app to create a new proxy server on every request (whether for the proxy or not), rather than just on startup. This led to a memory leak, and the Node process crashing.

This change creates it once, and uses the express-async-handler to get around triggering the no-misused-promises linting rule.

Refs #1636, https://github.com/chimurai/http-proxy-middleware/issues/108#issuecomment-1613754106, https://github.com/chimurai/http-proxy-middleware/tree/v3.0.0?tab=readme-ov-file#express-server-example, https://typescript-eslint.io/rules/no-misused-promises/