Fix the sandbox
There's a bug in the `@effect/sql-libsql` library where transactions don't work with a libSQL server. This has broken in the sandbox, but not locally where SQLite is used. This commit removes transactions temporarily to allow the sandbox to be used, as it's still feature-flagged and is solely for testing at the moment.

Refs #1973, b041ac188a89b27bad7400aa59e52a87ee16e57e, https://github.com/Effect-TS/effect/issues/3709