Mitigate problems using an in-memory database
The tests began to fail when re-enabling transactions as they're incompatible with in-memory SQLite databases. It should be possible to use `file::memory:?cache=shared` rather than `:memory:`, but this also doesn't entirely. To avoid spending more time on this problem, this commit uses temporary file-based databases in the tests.

Refs #1973, 3acf81478426d516a62e326aa4634fecba2a4623, https://github.com/tursodatabase/libsql-client-ts/issues/229, https://www.sqlite.org/inmemorydb.html