Pass the locale through to Express pages
This change avoids duplicating reading the cookie on the legacy pages by passing through the already-read value. This requires a change to our Express usage to allow for request-specific arguments. As we're not using Express to listen for requests, we've favoured this type-safe way (as opposed to Express's standard of modifying the request object).

Refs #2044