Allow HTML schemas to be made optional
I found some strange behaviour when making an HTML field optional: they would end up with the string 'null' rather than the type null. It turns out that using Schema.Any would cause, well, any value to become a string. This change is not right (the HTML type needs to be rethought), but it's less wrong than it was.

Refs #2060