Force CMS updates to appear
I've been unable to work out what's happening, but pages from the CMS haven't been updated. The revalidation request is made, but it uses the cached copy again rather than revalidating. I'm unsure if this is related to the recent changes to how content is read from the CMS, but it seems likely. If not, there might be a wider problem (as a lot of content uses this technique). To force through the changes, this commit removes the stale caching at the cost of a small performance/reliability regression.

Refs 201a9d41b2937183f2e873cc1acf84e790e25076, https://prereview.slack.com/archives/CR6B5804W/p1734390769283349?thread_ts=1734382300.226219&cid=CR6B5804W