Cater for Crowdin adding elements to the DOM
The Crowdin In-Context Editor adds an extra `span` element when there's whitespace at the beginning or end of its contents, even when HTML semantics collapses it. This change should allow all these extra `span`s to appear as blocks, to ensure consistency. (It looks odd when 2 paragraphs appear as blocks, then a 3rd appears as multiple lines.)

Refs #2116