Make Playwright more reliable by waiting for the server to start
I've noticed locally that Playwright often fails as it tries to use the browser before the server starts; I've now seen it happen in CI too.

This change waits for the health check to become available. I looked at using a library like 'wait-on', but it adds a lot of dependencies. Luckily, Effect makes this really straightforward.

Refs #2180