Avoid test failures
A HashSet inside an object inside an fp-ts Either isn't being compared properly: the order inside the internal array is different, but shouldn't matter. I looked into 'fixing' this by having Effect compare the two values properly, but I couldn't work it out. Instead, this change normalises the value before comparison. It's far from ideal, but this is code that should be migrated to Effect anyway.