Remove web-test-runner as it doesn’t work properly
As GitHub shows a red cross when an optional job fails, this change removes the use of web-test-runner entirely to avoid confusion. Due to our approach to client-side JavaScript and existing integration tests, it’s not worth the hassle. It also reduces the number of dependencies, including some that have vulnerabilities.

Refs #2017, b95fc362b34798e5d8544cb4064fbb35b68aadd9